ts-migration: convert some routes from js to ts (#1662)
* ts-migration: convert some routes from js to ts

* ts-migration: goals, issues and index page

* ts-migration: convert healchecks, progresses, pullrequests, staging and trading in typescript

* fix: issue route import name

* fix: goals route import name

* change notify route js to ts
