+++ routes/contributions.ts
@@ -1,6 +1,6 @@
-const express = require("express");
+import express from "express";
 const router = express.Router();
-const contributions = require("../controllers/contributions");
+import contributions from "../controllers/contributions";
 
 router.get("/:username", contributions.getUserContributions);
