--- routes/goals.js
@@ -1,8 +1,0 @@
-const express = require("express");
-const router = express.Router();
-const authenticate = require("../middlewares/authenticate");
-const goals = require("../controllers/goals");
-
-router.get("/token", authenticate, goals.getGoalSiteToken);
-
-module.exports = router;
