--- routes/notify.js
@@ -1,9 +1,0 @@
-const express = require("express");
-const router = express.Router();
-
-const authenticate = require("../middlewares/authenticate");
-const { notifyController } = require("../controllers/notify");
-const { notifyValidator } = require("../middlewares/validators/notify");
-
-router.post("/", authenticate, notifyValidator, notifyController);
-module.exports = router;
